pyproject.toml
[project]
name = "saucemall-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "pymysql>=1.1.2",
    "pytest>=8.4.2",
    "requests>=2.32.5",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[dependency-groups]
dev = [
    "allure-pytest>=2.15.3",
    "locust>=2.34.0",
    "playwright>=1.57.0",
]
